Move interceptor to shared folder
